More size_t.

+++ src/dynamic_buffer.h
@@ -29,9 +29,9 @@ struct dynamic_buffer_s
 typedef struct dynamic_buffer_s dynamic_buffer;
 
 /* Prototypes begin here */
-char *low_make_buf_space(INT32 space, dynamic_buffer *buf);
+char *low_make_buf_space(size_t space, dynamic_buffer *buf);
 void low_my_putchar(char b,dynamic_buffer *buf);
-void low_my_binary_strcat(const char *b,INT32 l,dynamic_buffer *buf);
+void low_my_binary_strcat(const char *b, size_t l, dynamic_buffer *buf);
 void debug_initialize_buf(dynamic_buffer *buf);
 void low_reinit_buf(dynamic_buffer *buf);
 void low_init_buf_with_string(string s, dynamic_buffer *buf);
